return_zone_approval/requirements.txt
# Slack Bot
slack-bolt>=1.18.0
slack-sdk>=3.21.0

# Email monitoring (Gmail API)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# AI parsing (Claude API)
anthropic>=0.18.0

# Web automation
playwright>=1.40.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
aiohttp>=3.9.0

# Database (for approval queue)
tinydb>=4.8.0
